refactor(activerecord): move Batches methods into relation/batches.ts and include Explain into Relation
Two RFC 0107 decomposition steps.

`findEach`, `findInBatches` and `inBatches` (265 lines of `relation.ts`)
implement `active_record/relation/batches.rb`; they now live in
`relation/batches.ts` alongside the private helpers they already used, mixed
into `Relation` via `include(Relation, Batches)` as
`relation.rb:68`'s `include ... Batches ...` does. Member order follows
batches.rb source order (find_each :36, find_in_batches :96, in_batches :167,
act_on_ignored_order :474). Bodies move unchanged except
`this._loaded` -> `this.loaded`, which is what batches.rb:212's `loaded?`
actually reads and which retires the call-set row the move surfaced. The
`in_batches`/`arel` baseline row moves with the method from
`relation.json` to `relation/batches.json`, reason unchanged.

`explain.ts` gains an `Explain` module object over its existing
`collectingQueriesForExplain` and `renderBind` definitions, so Rails' two
mixin sites both read one definition: `base.rb:294`'s `extend Explain`
(already wired in `base.ts`) and `relation.rb:68`'s `include ... Explain ...`,
which trails did not have at all — a `Relation` receiver now resolves both
names. No implementation is duplicated or changed.
