+++ scripts/api-compare/call-mismatches-exclude/activerecord/relation/batches.json
@@ -23,6 +23,13 @@
   {
     "package": "activerecord",
     "tsFile": "relation/batches.ts",
+    "rubyName": "in_batches",
+    "call": "arel",
+    "reason": "Confirmed equivalent (RFC 0047): the TS body builds the Arel manager via _buildArel/toArel (the build_arel port that arel() delegates to) rather than the memoized arel reader; surfaced when arel() gained its Rails-faithful aliases param (query_methods.rb:1594). See PR #4518."
+  },
+  {
+    "package": "activerecord",
+    "tsFile": "relation/batches.ts",
     "rubyName": "record_cursor_values",
     "call": "slice",
     "reason": "Baseline (RFC 0047): wide call-set flag seeded when the wide ratchet landed; bucket (b) equivalent or (c) noise pending per-cluster burndown review."
